// src/redux/types/ordersTypes.ts
export enum OrdersActionTypes {
  FETCH_ORDERS = 'FETCH_ORDERS',
  FETCH_ORDERS_SUCCESS = 'FETCH_ORDERS_SUCCESS',
  FETCH_ORDERS_ERROR = 'FETCH_ORDERS_ERROR',
  SET_ORDERS_PAGE = 'SET_ORDERS_PAGE',
  SET_ORDERS_PERIOD = 'SET_ORDERS_PERIOD',
  SELECT_PERIOD = 'SELECT_PERIOD',
  FETCH_ORDER_STATUS = 'FETCH_ORDER_STATUS',
  FETCH_ORDER_STATUS_SUCCESS = 'FETCH_ORDER_STATUS_SUCCESS',
  FETCH_ORDER_STATUS_ERROR = 'FETCH_ORDER_STATUS_ERROR',
}

export enum OrderStatuses {
  complete = '5e26a1f0099b810b946c5d8b',
  cancel = '5e26a1f5099b810b946c5d8c',
  restore = '5e26a191099b810b946c5d89',
}